==> script/3-final_visualization.Rmd <==
---
title: "Analyse_singlecell_para_okay"
author: "Malassigne_Victor"
date: "11/14/2022"
output: html_document
---

```{r, echo=FALSE, message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = T, fig.height=9, fig.width=9, fig.align="center")
library(Seurat)
library(dplyr)
library(khroma)
library(future) # parallelization in Seurat (NormalizeData() / ScaleData() / JackStraw() / FindMarkers() / FindIntegrationAnchors() / FindClusters() - if clustering over multiple resolutions)
library(clustree)
library(ggplot2)
library(openxlsx)
```

```{r, echo=FALSE, message=FALSE, warning=FALSE}
rm(list = ls())
plan("multiprocess", workers = 9) # activate parallelization
```

```{r, echo=FALSE, message=FALSE, warning=FALSE}
Singlet_norm <- readRDS("~/Documents/JM/singelcell_LY49/data/Singlet_norm.RDS")

my_palette = c("VV-B8R"= "#DDCC77",
               "Ly49p-CD8ab"= "#117733",
               "VV-Ly49p-CD8aa"= "#882255",
               "Ly49p-CD8aa"= "#88CCEE",
               "VV-Ly49p-CD8ab"= "#999933",
               "Naive"= "#AA4499",
               "Ly49n"= "#332288")
```

# PCA   
```{r, echo=FALSE, message=FALSE, warning=FALSE}
Singlet_norm <- RunPCA(Singlet_norm, features = VariableFeatures(object = Singlet_norm), assay = "SCT")
print(Singlet_norm[["pca"]], dims = 1:4, nfeatures = 10)

print("Contribution des gènes pour les PC")
VizDimLoadings(Singlet_norm, dims = 1:4, reduction = "pca", nfeatures = 10)

ElbowPlot(Singlet_norm, ndims = 50) +
  ggtitle(label = "Part de variabilité dans chaque PC")

print("Heatmaps avec les 20 gènes contribuants le plus dans la PC")
DimHeatmap(Singlet_norm, dims = 1:9, cells = 500, balanced = TRUE, nfeatures = 20)
DimHeatmap(Singlet_norm, dims = 10:18, cells = 500, balanced = TRUE, nfeatures = 20)

DimPlot(Singlet_norm,
        reduction = "pca",
        group.by = "HTO_maxID",
        pt.size = 1,
        cols = my_palette)+
  labs(color = "legend title")+
  ggtitle(label = "PCA")

DimPlot(Singlet_norm, 
        reduction = "pca",
        group.by = "HTO_maxID",
        dims = 3:4,
        pt.size = 1,
        cols = my_palette)+
  labs(color = "legend title")+
  ggtitle(label = "PCA")
```
  
Choose 11 CP for the continuation  

# UMAP  
```{r, echo=FALSE, message=FALSE, warning=FALSE}
Singlet_norm <- RunUMAP(Singlet_norm, dims = 1:11, assay = "SCT")
DimPlot(Singlet_norm, 
        reduction = "umap",
        group.by = "HTO_maxID", 
        pt.size = 1, 
        cols = my_palette,
        label = T) +
  labs(color = "Legend :") +
  ggtitle(label = paste0("UMAP 11 CP"))
```

# DE befor clusterization  
```{r, echo=FALSE, message=FALSE, warning=FALSE}
###### Change lvl ###### 
condition <- as.data.frame(Singlet_norm@meta.data[["HTO_maxID"]]) #creer un df de tous tes idividus
Singlet_norm[["orig.ident"]] <- condition$`Singlet_norm@meta.data[["HTO_maxID"]]`  #on passe la 1er colonne
Idents(Singlet_norm) <- condition$`Singlet_norm@meta.data[["HTO_maxID"]]`  # reviens a faire un levels

levels(Singlet_norm)

###### DE ###### 
Singlet_norm.markers_befor <- FindAllMarkers(Singlet_norm, 
                                       only.pos = F, 
                                       min.pct = 0.1, #  diff de 0.1 entre les deux conditions
                                       logfc.threshold = 0.5,
                                       min.diff.pct = 0.25) 

###### save xlsx ###### 
# write.xlsx(Singlet_norm.markers_befor, file = "/Users/victor/Documents/JM/singelcell_LY49/SC_Ly49_10X/results/DE_befor_clusterization.xlsx")

###### Visualization ###### 
Singlet_norm.markers_befor %>%
  group_by(cluster) %>%
  slice_max(n = 2, order_by = avg_log2FC)

Singlet_norm.markers_befor %>%
  group_by(cluster) %>%
  top_n(n = 10, wt = avg_log2FC) -> top10

DoHeatmap(Singlet_norm, features = top10$gene) + 
  NoLegend() +
  ggtitle(label = "Heatmaps du DE entre les différentes population")
```

# Clustree and clusterization  
```{r, echo=FALSE, message=FALSE, warning=FALSE}
Singlet_norm <- FindNeighbors(Singlet_norm, dims = 1:11, verbose = FALSE)
Singlet_norm <- FindClusters(Singlet_norm, 
                             resolution = seq(0,1.5,0.1))
clustree(Singlet_norm, prefix = "SCT_snn_res.") + 
  ggtitle(label = "Clustree avec 11 CP")

Singlet_norm <- FindClusters(Singlet_norm, resolution = 0.4)
Singlet_norm <- RunUMAP(Singlet_norm, dims = 1:11, assay = "SCT")
DimPlot(Singlet_norm, 
        reduction = "umap", 
        pt.size = 1,
        label = T)+
  ggtitle(label = paste0( "UMAP 11 CP et resolution : 0.4"))
```

# DE after clusterization  
```{r, echo=FALSE, message=FALSE, warning=FALSE}
levels(Singlet_norm)

###### DE ###### 
Singlet_norm.markers_after <- FindAllMarkers(Singlet_norm, 
                                       only.pos = F, 
                                       min.pct = 0.1, 
                                       logfc.threshold = 0.5,
                                       min.diff.pct = 0.25)

###### save xlsx ###### 
# write.xlsx(Singlet_norm.markers_after, file = "/Users/victor/Documents/JM/singelcell_LY49/SC_Ly49_10X/results/DE_after_clusterization_reso04.xlsx")

###### Visualization ###### 
Singlet_norm.markers_after %>%
  group_by(cluster) %>%
  slice_max(n = 2, order_by = avg_log2FC)

Singlet_norm.markers_after %>%
  group_by(cluster) %>%
  top_n(n = 10, wt = avg_log2FC) -> top10

DoHeatmap(Singlet_norm, features = top10$gene) + 
  NoLegend() +
  ggtitle(label = "Heatmaps du DE entre les différents clusters")
```

# Save seurat objet
```{r, echo=FALSE, message=FALSE, warning=FALSE}
# saveRDS(Singlet_norm, "/Users/victor/Documents/JM/singelcell_LY49/data/Singlet_norm_3.RDS")
```
